renames DashboardModal to CreateListingModal
